Fix issue with undefined label color for prediction
