+++ geti_sdk/rest_clients/prediction_client.py
@@ -723,4 +723,5 @@ class PredictionClient:
             contenttype=contenttype,
             data=data,
         )
-        return PredictionRESTConverter.from_dict(response)
+        prediction = PredictionRESTConverter.from_dict(response)
+        prediction.resolve_label_names_and_colors(labels=self._labels)
